auth_test: permissions_caching_test: improve timing validation
We're seeing suprious failures like the following:
AssertionError: Expecting query to raise an exception, but nothing was raised.

This patch relaxes timing validation by taking `grant_time`
before the GRANT query is executed and also comparing `delta`
to be great than *or equal* to 2 seconds, so that we retry
if the SELECT query succeeded with no excepton unexpectedly
and we suspect the permissions cache expired.

Fixes #1148

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>