hintedhandoff_removenode_test: improve log watching
Change the message we're looking for to
  hints_manager - ep_manager(<node_addr>)::sender: exiting
since it's always logged, even when there are no hints to drain.
(until we have a better indication in scylla)

Also, watch the logs of both node1 and node2 as node2 might still
be draining while node1 has finished, causing the test for
empty hints dirs to fail.

Note: we're going to add a more indicative message to scylla
to be supported in a different patch so that this patch
can be backported with no change to scylla.

Refs #5354

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>