compaction_additional_test: compaction_is_started_on_boot_test: support resharding compaction
IncrementalCompactionStrategy performs resharding compaction on the sstables
that this test renames, therefore the number of sstables will not change.

To support that, change the expected regular expression to detect
both regular compaction and rehasrding compaction.

That said, that may happen with other compaction strategies in case
they add the same number of sstables they happen to delete.

To deal with this scenario, rather than comparing the number
of sstables in the directory, compare the actual sorted list
of sstable files to make sure it changed.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>