hinted_handoff_additional_test.py: hintedhandoff_removenode_test(): watch for a log message that ensures the manager::drain_for() is complete
The message we were watching before denoted the beginning of the draining
and we need to be looking for the one denoting the end of it.

In addition peviously the test was not watching for corresponding messages on both nodes, on
all shards and from all hints managers.

This patch fixes all that.

Fixes scylladb/scylla#5354

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>