in-memory: Refactor memory per key ratio
Part of the in memory validation test depend on an empirically
determined factor, which is used in the code as a magic number.
This patch defines it as a member of the InMemory tester class.
And elaborate a little bit more about the factor and the assumptions
that it implies.