in-memory: Readjust the mem per key factor
The previous factor was not computed for mc format, this causes
to the test to fail since now mc format is on by default.
This patch adjusts the factor to include the mc format case.
It was computed using the output from the assertion in the
failed test. The actual computation result was 33 and the
parameter was set to 30 to leave a little head room since
the linear dependant of the memory usage uppon the number
of keys is only an approximation, especially in mc format.