make _cleanup_cluster (and _force_clean) available for class methods
Current version that does not support reusing cluster between tests.
The clusters are cleared in tearDown / cleanUpCluster.

We are adding the ability to reuse a cluster between tests. As such
reused clusters need to be preserved after tearDown / cleanupUpCluster
calls. The preserved cluster will be cleared when the tearDownClass will
be called. As such _cleanup_cluster must be made availble as a class method.

Signed-off-by: Shlomi Livne <shlomi@scylladb.com>