add support for reusing a cluster between tests of the same test class
- Reusing the cluster between different tests of the same test class can
only be done by running those tests sequentially one after the other on
the same cluster.
- Reuse of clusters is controled via env REUSE_CLUSTER
- If a test class is to reuse clusters it must inherit from
TesterReuseCluster instead of Tester.
- When REUSE_CLUSTER is True then _multiprocess_can_split_ will be set to
False for these tests forcing all the tests in the class to be run on the
same thread, allowing them to actually reuse the cluster
- When REUSE_CLUSTER is False then _multiprocess_can_split_ will be set
to True for these tests allowing running the tests in parallel on
different threads (without reusing the cluster)
- PRESERVED_CLUSTER is used to store a cluster following a sucessfull
test completion.
- If a test fails the cluster is cleared
- On test setup if PRESERVED_CLUSTER is avilable it will be used
- After running the last test of the class - the PRESERVED_CLUSTER will
be cleared

Unfortunatly multiprocess evaluates the test class _multiprocess_can_split_
early on (prior to calling setUpClass) as such there is no way to use an
@attr only and a such subclass must be used.

Signed-off-by: Shlomi Livne <shlomi@scylladb.com>