lwt: add prepared statement tests for LWT
1. Introduce `TestCQL.lwt_update_primitive_types_test`,
which iterates over every primitive CQL type, creates a test table
with a column of desired type, then executes two consecutive CAS
statements to verify that one succeeds while another fails to apply.

This test executes all variants of CAS conditions, i.e.
 * value=:v
 * value in (:v)
 * value in :v

Common arguments to supply to individual sub-test cases are as follows:
 * zero values
 * empty values (e.g. empty strings)
 * min/max values for the given type
 * some random value in-between the limits

Currently there is one thing that is not working (suspecting a bug):
passing a tuple containing null value to the CAS condition of the
form 'value in :v', so these are removed from the test case until
properly investigated.

2. Introduce `TestCQL.lwt_update_primitive_types_listlike_test`.
This test case is equivalent to the former except that all arguments
adjusted to work with lists and sets.

Tested collection types (both frozen and non-frozen variants):
 * list
 * set

For each of the mentioned above there are the following sub-cases
being executed:
 * value=:v
 * value in (:v)
 * value in :v

For lists also the following combinations are enabled (because
sets don't support any element access syntax for now):
 * value[:i]=:v
 * value[:i] in (:v)
 * value[:i] in :v

Again, all the value combinations are tested except for null values.

3. Introduce `TestCQL.null_value_tuple_test`

This test exhibits a difference between C* and Scylla behavior
(possibly a bug) regarding handling of 'value in :v' for null value
tuples in LWT conditions (in prepared statements).

It fails on Scylla and runs fine on C*.

Need to investigate further.

Signed-off-by: Pavel Solodovnikov <pa.solodovnikov@scylladb.com>