dtest docker: add support for syslogd logging
Docker containers lack the syslogd interface, in order to
have a syslog socket, docker needs to have rsyslog with
a suitable configuration. This commit adds rsyslog support
for the all future dockers that will be built.
It adds the installation and configuration necessary for
rsyslog to work under docker and to expose the standard
/dev/log socket.