replace_address_test: _replace_node_test: set no_wait=True when expected
Fixes #1297

Currently the replaced node starts listenting for CQL but this is a bug,
see https://github.com/scylladb/scylla/issues/5523, so do not wait
for binary proto in this case, similar to `replace_first_boot_test`.

Once scylladb/scylla#5523 is fixed we should update the test
to verify that the restarted node doesn't start serving.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>