replace_address_test: set wait_other_notice=False when expected
When replacing a node's address we expect it to fail to join the cluster
if restarted.  Since ee790f2484ef8d7defc27f87675771059d101550, the default
falgs are to wait for the node the start listening for CQL and for
others to notice it started - causing replace_shutdown_node_test to fail
since it restarts nodes with the default flags. Therefore, explicitly
call node.start(wait_for_binary_proto=True, wait_other_notice=False).
Similarly, do the same in replace_first_boot_test (that used `no_wait=True`
that doesn't even wait for the node to start listening for CQL).

Fixes #1297

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>