compaction_additional_test: remove _multiprocess_can_split_
Added by 6f866dc9 to prevent compact_data_by_time_window_test
to run in parallel with other tests. However, this flag affects
only other tests in this module, not in other modules, so
in practice it was running in parallel with other tests and
is passing happily.

The down side of this flag is that it is impossible to run
a single test case since nosetests will run all tests in the
module when this flag is set True.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>