lwt: split lwt_nested_collections_test into two tests
This patch splits one combined lwt_nested_collections_test into
two smaller tests, each one testing one variant of nested
collections:
 * lwt_nested_collections_list_set_test for list<frozen<set<T>>>
 * lwt_nested_collections_set_list_test for set<frozen<list<T>>>

This is more maintainable and in case some checks fail for
one of the collections, the other will still run and will not
be interrupted by a failure.

An dalso we are getting the knowledge of which one precisely
failed without even looking into the test logs.

Signed-off-by: Pavel Solodovnikov <pa.solodovnikov@scylladb.com>