auth tests: Fix use attempt of deprecated table
The tests contained in auth_tests.py used nodetool to
indirectly query system_auth.users which doesn't exist
anymore. The table name has been changed to represents
the move to the concept of roles and is now called "roles"
this patch changes the tests to use the new table name.

Tests: ran locally all of the tests in the auth_tests.py file
and made sure the pass.

Fixes #1331