replace_address_test.py: Fix replace_first_boot_test
```
222         # check that restarting node 3 doesn't work
223         # FIXME: https://github.com/scylladb/scylla/issues/5523 is fixed
224         # need to verify that the node doesn't start listening
225         debug("Try to restart node 3 (should fail)")
226         node3.start(no_wait=True)
227         checkCollision = node1.grep_log("between
	    .*"+self.cluster.get_node_ip(3)+" and
	    .*"+self.cluster.get_node_ip(4)+";
	    .*"+self.cluster.get_node_ip(4)+" is the new owner")
228         debug(checkCollision)
229         self.assertEqual(len(checkCollision), 1)
230
231         # restart node4 (if error's might have to change num_tokens)
232         node4.stop(gently=False)
233         node4.start(wait_for_binary_proto=True, wait_other_notice=False)
234
235         debug("Verifying querying works again.")
236         finalData = list(session.execute(query))
237         self.assertCountEqual(initialData, finalData)
238
239         # we redo this check because restarting node should not
	    result in tokens being moved again, ie number should be same
240         debug("Verifying tokens migrated sucessfully")
241         movedTokensList = node4.grep_log("Token .* changing
	    ownership from .*"+self.cluster.get_node_ip(3)+" to
	    .*"+self.cluster.get_node_ip(4))
242         debug(movedTokensList[0])
243         self.assertEqual(len(movedTokensList), numNodes)
```

The check at line 227 will pass when n4 is finished replacing. It has
nothing to do with the restart of the replaced node here. The restart of
replaced node n3 will not fail at all.

At line 232, we restart n4 immediately, note at line 226, we did not
wait fro n3's restart. If we really want to see what happens if n3
restarts, we should wait for it to finish restarting.

The check at line 243 is problematic, because nothing prevents we get
the log changing ownership from n3 to n4 again, after restart of n3 and
n4.

After n3's restart at line 226, n3 will be in gossip, when n4 restarts,
if n3's state is present at n4, n4 then goes to normal, we sill see n4
logs that ownership is changed from n3 to n4.

In this patch, we stop restarting n3. We can re-add the restarting of n3
when we support detecting replaced node to be added in the cluster and
reject the replaced node.