materialized_views_test: revert to 1s wait time in eventually()
Making `eventually()` use an exponential timeout was a mistake,
because if a correct number of rows happens to be propagated to a view
in its 34th second, the test will still wait for another ~30 seconds
without a good reason. Polling once per second was the way this test
worked before, and it's in every way superior to the exponential one,
so it's hereby reverted. The original patch was created in order to
increase the timeout from ~30s to ~60s, so the new default number
of retries is set to 64.