audit_tests: add expected errors waited for in watch_logs_for
Two tests in audit set wrong settings and waited on the log
for the exceptions. The problem was that once the error was
observed the test failed in the general post checks where
the same error were identified as unexpected errors.
This commit adds those expected errors to the relevant
tests.