dtest: default to the random CLUSTER_ID_ALLOCATOR
Otherwise, we can't catch issues with tests or code that uses the localhost
address by mistake, like the one we current see with `cdc_tests.TestCdc.simple_cdc_test`.

Also, this is the way the tests typically run no in jenkins and we should run them
in a similar way when run locally.

If needed, CLUSTER_ID_ALLOCATOR=single or NOSE_PROCESSES=0 can be used to select
the `SingleClusterIdAllocator` that was used by default when NOSE_PROCESSES wasn't
set or was set to 0.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>