cdc_test: fix sequential population of cluster
In order for CDC to work correctly, the nodes of the cluster need to be
added sequentially - the new node can be added only after the previous
one joined the cluster.

The function that was responsible for creation of the cluster in the way
described above was incorrect, as it didn't cause the cluster
configuration to be written. This would cause the first node try
gossiping with 127.0.0.1 which is the default seed. Due to the fact that
dtests may set a different address for the first node, setting up the
first node could fail as it would not be able to communicate with
configured seed.

Now, populate_sequentially initializes the first node using
cluster.populate(1), and adds remaining nodes in the old way.