repair_additional_test: repair_option_pr_multi_dc_test: increase num_keys
Double num_keys and increase the allowed range.

This should make the test less flaky by increasing the key space from
which the primary ranges are taken when repairing a single so it
will deviate less around the expected 1/6'th of the keys written per node
on average.

Fixes scylladb/scylla#6063

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>