nodetool_additional_test: concurrent_drain_test: clarify node_to_drain
Currently the test used 0-based index of the node to be drained (`1`,
which means node2).  Change this to use 1-based numbering and adjust it
in self.drain that for calling self.get_node.

node_to_drain will be used in the next patch to calculate the
expected node ip address in the cassandra-stress expected_errors.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>