lwt_test.py: fix flaky metrics and read_round_optimization tests
Some of the metrics can be incremented twice if LWT is
bounced to a different shard due to gh-5860

Rewrite read_round_optimization_test to remove flakiness.

Update 100 rows and check all 100 updates failed a read round
optimization. If the failure is flaky, the metric will be wrong.
Use QUORUM writes in non-paxos statement to make sure that
a different row copy reaches all nodes (the default is ONE).
Do not drop the table, it won't work predictably with a node down,
and it is too long to start this node, we're dropping the entire
cluster anyway.

Remove tracing, since it hides possible races.

Closes #6125