twcs streaming: unflake flakiness of flaky test
test_streaming_sstables_with_several_timewindows fails sometimes.
The full discussion as to why is at https://github.com/scylladb/scylla/issues/6162

The reason for that is that the test inserts 6,000 keys and expects
to verify that the data streamed obeys the twcs properties.

However, because the RF is one, it might be that all the data happens
to belong to just the first node anyway, in which case no data is
transferred.

With RF=2 we can test what we want just as well, and then we have
the guarantee of data transfer. Move this to RF=2.

Signed-off-by: Glauber Costa <glauber@scylladb.com>