topology_test: decommissioned_node_cant_rejoin_test: simplify use of watch_log_for
It can simply be used after start(no_wait=True), no need for the try/except
song and dance and no need to grep_log_for_errors after it since that's
what watch_log_for is doing, for a specific error.

Also, allow more time for the node to stop after seeing the `rejoin_err`
message in the log.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>