sla tests: fix expected error type in "non-existing" tests
The error that was returned from Scylla when trying to operate
on a non existing role or service level was invalid. Scylla
returned error code 0000 (Server Error) while in fact the query
was just invalid which corresponds to code 2200.
This commit changes the expected error to the correct one.

Testing: Ran the tests with a fixed version of Scylla which
returns the correct error codes.