test(lwt_banking_test): emulating bank transaction
The idea of the banking test is in emulating a series of bank money
transfers between accounts.  We try to leverage LWT here to make our
transaction more error proof.

The core of the test is a transaction and it contains several steps:

1) Add new entry to `transactions` table with status `in progress`;
2) Add new entry to `transactions_in_progress`;
3) Mark source and destination accounts as involved to the transaction
        a) in case of the same partition we can do it one batch
        b) for different partitions we need to split it to two updates
4) Get balances of the accounts;
5) Change balances of the accounts and mark them as not involved to any
transaction (same about partitions as in step 3);
6) Delete the entry from `transactions_in_progress` table;
7) Mark the transaction as completed.

After every transaction sum of all balances should be the same as before
the transaction.