Adding a function to tools.py to easily create a compatible table with cassandra-stress
create_stress_compatible_table it a function that allows to create the keyspace and the
table that are compatible with cassandra-stress.
This is helpful when wanting to run the stress with some values other than the default.
For example, creating the table with dclocal_read_repair set to 0.1 or speculative_retry
set to 'NONE'.

pre-create c-s table with dclocal_read_repair_chance set to 0.1