cdc_types_test: make valid class identifiers
Characters like '[<>, ]' are invalid in identifiers
so they shouldn't be used to generate `cls_name`, as
the latter cannot be used to run a particular test.

We also see that in dtest-debug-random that selects a random
subset of the tests and hits issues with quoting names
containing redirection symbols, and then even if these are
well-quoted, the individual tests cannot be adressed
when they contain invalid identifier characters.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>