run_test.sh: derive scylla sibling dirs from scylla product
Auto-detect SCYLLA_PRODUCT from CASSANDRA_DIR, use
it to derive the respective jmx and tools-java sibling directories
to select either scylla-jmx or scylla-enterprise-jmx, depending
on whether SCYLLA_ROOT_DIR is `scylla` or `scylla-enterprise`.

Although typically scylla-enterprise-{dtest,ccm} should be used to
to test scylla-enterprise, it is possible to test scylla-enterprise
from scylla-dtest. In that case, the *-ccm sibling dir should be
derived from the dtest directory, not the scylla directory.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>