internode_ssl_test: ignore rpc errors when reloading certificates
rpc errors were logged in scylla in INFO severity historically,
but that was changed to ERROR and addressed globally in #1421.

This change adds particular error messages that are expected
in this test to `ignore_log_patterns`.

It fixes spurious failures as seen in:
https://jenkins.scylladb.com/view/master/job/scylla-master/job/dtest-release/483/testReport/junit/internode_ssl_test/TestInternodeSSL/putget_with_reloaded_certificates_test/
https://jenkins.scylladb.com/view/master/job/scylla-master/job/dtest-release/492/testReport/junit/internode_ssl_test/TestInternodeSSL/putget_with_reloaded_certificates_test/
```
Unexpected errors found: [('node2', ['ERROR 2020-05-23 08:48:22,573 [shard 0] rpc - client 127.0.0.1:61434: server connection dropped: The TLS connection was non-properly terminated.']), ('node3', ['ERROR 2020-05-23 08:48:22,572 [shard 0] rpc - client 127.0.41.2:7001: client connection dropped: The certificate is NOT trusted. The certificate issuer is unknown. '])]
```

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>