test(manager): snapshots of a fail backup are deleted upon rerun
Added a manager test that generate data in the cluster, create a backup task and cause it to fail.
Afterwards, the test modifies the data, and rerun the backup task. At this point, the tast makes sure
that the snapshots from the failed backup are deleted at the baginning of the rerun