migration_test.py: derive format from test source dir, not cf_dir
This test assumes that the upload process will not change the format
of the SSTables. With this assumption, it scans the table directory
(cf_dir) to find out what the before format is.

This assumption is not valid, and even today can break in some cases.
With an upcoming patchset of mine it will break in all cases.

The solution is to grab the format from the test source directory.

Signed-off-by: Glauber Costa <glauber@scylladb.com>