fix(replace_address_test.py): tokens check were failing
As reported on scylladb/scylla#6437 the verification
of replace_address_tests were failing when searching
in node logs.
As a better solution, now getting from the original
node the list of tokens and comparing with the node
that is replacing it with the list of tokens (both
lists must be the same). Getting the token lists
running `nodetool ring`.