LWT Banking load test
Ported test from lightest/go.

Banking test

The idea of banking test is in emulating a series of bank money transfers
between accounts. A typical bank has transfers between corporate accounts,
personal accounts, and corporate-to-personal accounts (e.g. salary payout).
Transfers can be internal and external, i.e. involve a third party bank.

The key idea which makes the test useful for data validation, even though it
doesn’t use an oracle, is that no amount of money transfers can change the
total balance for accounts managed by the banks. The test thus consists of 3
major steps:

1. Data generation.
Bank accounts are batch loaded, including some initial balances. A personal
account differs from a corporate account only by a lower initial balance
amount. The total balance is calculated and recorded as a canonical /expected
result. This doesn’t have to be LWT workload, the data can be generated in
batches.
2. A series of money transfers between accounts is performed. E.g. a salary
pay out at the end of month for a corporate account, and peer-to-peer
transfers for a personal account. The transfers will compete against each
other since they may use the same source or destination account. The length
of this step, the level of randomness, as well as used nemeses can be
arbitrary and does not impact the validation part of the test. The key
property of this test, however, is that it should also emulate faulty clients
- i.e. clients which experience networking issues or fail midway a transfer
operation.
3. The account balances are scanned and the total balance must not change.

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>