tools: don't skip in @since decorator
Just print a debug message at this point since
scylla version cannot be used for this at the moment.
It is either a constant '3.0' at the moment
when running from source, or the SCYLLA-VERSION-FILE
if running from a relocatable package, but this isn't
helpful for comparing to an open source base version
like '3.3'.

Test: paxos_tests:TestPaxos.test_topology_change_in_presence_of_down_node

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>