nodetool_additional_test: getendpoints should use value
There is a problem with the current implementation of the helper method
getendpoints_test.
Instead of using the supplied value, it uses the string `value` there
was a bug in the get endpoints implementation which cause the test bug
to pass.
Now that Scylla's code was fixed the test fails.

This patch changes the implementation to use the actual value.

Fixes scylladb/scylla#7145

Signed-off-by: Amnon Heiman <amnon@scylladb.com>