paxos_tests: disable schema_mismatch_mv_test due to scylladb/scylla#7225
The test fails occasionally when using a debug scylla build on
Jenkins because the test depends on obsolete schema versions
being present in the local schema registry cache, but
when the node is too slow, the entries are evicted too soon from
the cache (the timeout is 1 seconds and not configurable at the
moment).

Skip the test until the underlying issue is resolved and the test
is adjusted accordingly.

Signed-off-by: Pavel Solodovnikov <pa.solodovnikov@scylladb.com>