alternator_utils.py: don't rely on old version of mypy_boto3_dynamodb
In alternator_utils.py, we relied on mypy_boto3_dynamodb containing a
certain definition, "AttributeValueTypeDef". However, in more recent
versions (I tested 1.14.62.0) it does not, so the Alternator tests
no longer work.

There is no reason to rely on this internal implementation details
of the mypy_boto3_dynamodb which can (and indeed, did) go away,
we can easily define it ourselves according to the official DynamoDB
API. So this is what this patch is doing.

With this patch, the Alternator tests work correctly with newer versions of
mypy_boto3_dynamodb.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>