in-memory tests: fix expected failure for out of space tests
The in memory expect failure attribute wasn't adjusted to the
new mechanism of log filtering pattern. In addition one of the
out of memory tests wasn't expecting the right exception type.
This commit fixes the "more_data_than_mem_available" family of
tests.

Tests: Ran the tests locally before and after the fix.