LDAP container: Dump server log when an error occurs
We see that sometimes, some LDAP server behaviours are
environment dependant (i.e Jenkins) . When something goes
wrong during LDAP operations, the server sometimes returns
an error which is translated to an Exception which fails
the test (as it should), however, we lack the information
from the server when analyzing the failure. Here we add
code that when those error happens, dumps the LDAP server's
logs as debug information before failing.