hintedhandoff_additional_test: improve hh_enabled_value sanitization
The assert introduced in 1e221d68e55223d855e31f475478eb42713d549f
is too strict as scylla `--hinted-handoff-enabled` option can
be either boolean (true|false|1|0), or a comma separated list
of data center names for which to enable hintedhandoffs.

Thus syntax is used in `hintedhandoff_rebalance_test` that
passes a non-existing dc name ('dont_send_hints') on purpose.

The assertion failure can be seen in
https://jenkins.scylladb.com/view/master/job/scylla-master/job/dtest-release/648/testReport/hintedhandoff_additional_test/TestHintedHandoff/hintedhandoff_rebalance_test/
```
Traceback (most recent call last):
  File "/usr/lib64/python3.7/unittest/case.py", line 60, in testPartExecutor
    yield
  File "/usr/lib64/python3.7/unittest/case.py", line 645, in run
    testMethod()
  File "/jenkins/workspace/scylla-master/dtest-release/scylla-dtest/hintedhandoff_additional_test.py", line 57, in hintedhandoff_rebalance_test
    self.__start_all([node1, node2, node3], hh_enabled_value='dont_send_hints', extra_jvm_args=['--smp', '3'])
  File "/jenkins/workspace/scylla-master/dtest-release/scylla-dtest/hintedhandoff_additional_test.py", line 520, in __start_all
    jvm_args=self.__jvm_args(hh_enabled_value=hh_enabled_value) + extra_jvm_args)
  File "/jenkins/workspace/scylla-master/dtest-release/scylla-dtest/hintedhandoff_additional_test.py", line 475, in __jvm_args
    assert hh_enabled_value == 'true' or hh_enabled_value == 'false'
```

This change introduces `__sanitize_hh_enabled_value` that
verifies the value and returns a sanitized string.

We call it both in `__jvm_args` and in `__start_all`
for printing a sensible debug string.
Previously, it would falsely print "enabled" even if
the value was "false", as a string holding "false" is
implicitly converted to a True boolean value since it
is not an empty.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>