hintedhandoff_additional_test: hintedhandoff_rebalance_test: adjust op_cnt
1. Inserting 1000000 keys needlessly takes too long in release mode.
Reduce it to 100000 as it creates more than enough hint logs for rebalancing
and takes only about a minute (on my machine) in sev mode.
2. Inserting 10000 keys in debug mode generates only a single hints log
file, so the rebalancing code isn't exercised. Reduce it to 25000
that generates over 10 hint logs so it exercises the rebalancing logic and takes
only about 3 minutes (on my machine).

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>