paxos_tests: provide schema_mismatch tests both for in-memory and persisted cache
Prior to the patch all `schema_mismatch_*` tests were enabled to
test only the case when in-memory schema registry cache is used
to recover obsolete schema version during a failed LWT operation.

Now these tests are renamed to have a "cache" suffix in the name
and also there are "no_cache" tests which will restart the node
in order to clear in-memory cache so that the node will start
looking for a persisted column mapping to resolve obsolete schema.

Refs: scylladb/scylla#6074
Tests: paxos_tests.py (TestPaxos.schema_registry_*_test)

Signed-off-by: Pavel Solodovnikov <pa.solodovnikov@scylladb.com>