ldap docker: Try to recover ldap connection if terminated
This change tries to recover a connection if it is was terminated
by the server. The reasons for why it happens are still unknown but
we see it happening in practice.
One more thing this commit does is to raise the LDAP server logging
level so it might reveal more if our test fails while interacting
with the LDAP server.