ldap docker: Some minor fixes to the create connection function
The LDAP create connection function can be retried up to 5 times,
however, some of the objects might already have been created so we
skip creating them in the following invocations.
One more tweak is that we don't call the open method of the connection
since the bing method will call it if needed.