ldap docker: Create connection only after ldap startup
It turns out that the docker image that we use start LDAP
twice, at the first time it start it up for configuration,
then it kills it and restart it again properly.
It can happen on a slow enough machine that the connection
is create on the instance of ldap that was started for configuration
sake. This is bad since at some point the instance will be killed
and the connection will become invalid.
This commit add logic that waits for the ldap server's startup before
opening a connection to the container.

Fixes #1797