ldap docker: Remove insane dbugging and logging level
We have increased the logging level of the LDAP server
very high to debug a problem where an ldap connection
goes stale after it has been established. We have also
added some recovery code that is not necessary anymore.
Now, we can lower the level and remove the excess recovery
since we found the problem :)