cql_tests: add test_query_failed_when_node_is_down tests
Test that a select query with consistency_level=QUOROM
returns an Unavailable error properly when one out of two nodes is DOWN.

We test two scenarios:
1. When one of the nodes is stopped gently, so the remaining node
is notified that the other node is down.  In this case the select
query is expected to fail immediately.
2. In the other case, we kill the node using stop(gently=False).
The select query is expected to time out on first attempt,
and the be retried and fail as Unavaiable as the node should be marked
DOWN by gossip.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>