replace_address_test: replace_first_boot_test: reduce range_request_timeout_in_ms
Although accepting OperationTimedOut exception as done
in 679f93909943304202173276e79036102c0d89d0
will help the test pass, it is an indication that the client
timed out the query before the server.

It'd be better to reduce the configured range_request_timeout_in_ms
such that the server will time it out before the 30 seconds
session.execute timeout, and verify that we get Unavailable
exception as we used to.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>