cql_tests: test_query_failed_when_node_is_down: verify immediate timeout
Verify that the query with cl=QUORUM times out immediately
within 1 second when the node is stopped gently, or after
a number retries, crossing 10 seconds (plus 1 second grace),
if the node is killed forcibly.

The 10 seconds minimum timeout is presumed to be
coming from the time to take gossip to mark
the node as down.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>