materialized_views_test: _do_resharding_test: ignore seastar::broken_promise with interrupt_resharding
Comapction fails with
```
compaction_manager - resharding failed: seastar::broken_promise (broken promise)
```
for now, until the reader is properly aborted and closed.

Since it's benign (certianly for the purpose of this test)
that compaction reports this error after it was aborted,
just ignore this error until the scylla issue is resolved.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>