ttl_test: deflake insert_ttl_has_priority_on_defaut_ttl_test
There as a period around the estimated expiration
time where we can't be absolutely sure if the data has
expired or not.

Therefore, start re-examining it about a second
before it's expected to expire and repeat every 0.1
seconds until 5 seconds from the timestamp before
the INSERT query, rounded to whole seconds (as is
Scylla gc_clock resolution).  This is the last time
we can be sure the data should not expire.

Then, sleep 5 seconds, from the timestamp _after_
the data was inserted, rounded down to a whole second.
This is the first time where we can be sure that the data
has indedd expired, and verify it is indeed not retrieved.

Fixes #1988

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>