test(manager): snapshot files of a deleted backup are purged
New in manager 2.4
As of 2.4, during a purge, the manager will remove all of the snapshot files
of a deleted task from the bucket, if the age of the manifest is over a month old.
The test runs and deletes a backup task, in order to create orphan files in
the bucket, and alters the date in the name of the manifest to fool the manager
to think that the manifest is over a month old, and therefore all of the snapshot
files the manifest points to will be removed from the bucket.