refactor(nameless_index_test): Change the number of nodes in test
This test sometimes fails durin rolling-upgrade test when it's being
run on cluster that is larger than one node.
There was a discussion about it as part of scylla issue:
https://github.com/scylladb/scylla/issues/8207.

It's better to find the issue here and not as part of rolling ugrade.

(cherry picked from commit b88bad082fabb0eb119db4192d3f53b4782bb803)