cql_tests: AbortedQueriesTester: adjust tests for scylla
The tests were skipped with @require('7392')

As per b8d7aab287116fe56c4b4552070cca50c74d1500,
it refers to CASSANDRA-7392, not to a scylla github issue.

https://issues.apache.org/jira/browse/CASSANDRA-7392
is resolved in 3.2 so the @require is not relevant anymore.
However the tests contain Cassandra specific configuration.

Adjust the tests for triggering timeout on scylla based
on the cluster type and reenable them.

DTest: cql_tests.py:AbortedQueriesTester(dev, debug)
Signed-off-by: Benny Halevy <bhalevy@scylladb.com>