test(manager): running a backup with purge-only param
New in manager 2.4
With the purge-only param, the backup task will execute a purge
on the specified bucket as expected, but will not execute a normal backup afterwards.
The test confirms that, while the purge indeed ran on the bucket,
no backup actually ran. That test does that by making sure that
after the task's completion there are no snapshot files of the cluster in the bucket.