test(manager): making sure backup validate task deletes orphan files
New in manager 2.4
In Scylla Manager terminology, an orphan file is a file that exists in the bucket
even though it is not documented in the backup manifests.
When the user starts a backup validate task, the task identifies any orphan files in the bucket,
and when the delete-orphaned-files is used, the task automatically deletes them.
The new test, test_backup_validate_delete_orphaned_files, makes sure those files
are, indeed, deleted.