test(manager): Error cql status when ssl is activated
New in manager 2.5

In the latest manager update, a new cql status was introduced to the
cluster status table - ERROR. That status appears when, due to an internal
issue, the manager ('s agent) could not communicate with the node with cql.

The new test, test_error_cql_status_when_ssl_is_activated, creates this kind
of scenario: it starts a normal, default cluster and adds it to the manager,
and afterwards activates ssl encryption for that cluster without updating
the manager, thus causing an internal error when the manager tries to
check the cql status of the cluster.

In addition, the get_hosts_health function and HostHealth object were modified
to locate and keep the new error messages that can appear after the status table.

*pytest move fix: The secondary cluster now starts without a manager server