test(migration_test): added tests to cover new feature 'load and stream' (#2245)
* test(migration_test): added tests to cover new feature 'load and stream'

Issue:
https://github.com/scylladb/scylla/commit/df3ef800c20c60d7929ffa600649793aa6c73064

Task:
https://trello.com/c/nwktmFqq/3478-dtest-7831-the-load-and-stream-feature-has-been-merged-
this-allows-nodetool-refresh-to-split-loaded-sstables-into-owned-and-non

This feature extends the nodetool refresh to allow loading arbitrary sstables
that do not belong to a node into the cluster. It loads the sstables from disk
and calculates the owning nodes of the data and streams to the owners
automatically.

Added to tests:
1. load sstables from 4-nodes cluster to 2-nodes cluster
2. load sstables from 2-nodes cluster to 4-nodes cluster
3. primary_replica_only parameter test
4. table with frozen<UDT> primary key
5. table with 2 columns partition key

* test(load and stream): data for the tests of 'load and stream' feature

(cherry picked from commit 3031d69d48a103943b5bfa5ae646f0d342cae775)